+++ src/commands/game_c4/mod.rs
@@ -4,3 +4,4 @@ pub use discord_hooks::ConnectFourDiscord;
 mod board;
 mod c4;
 mod discord_hooks;
+mod discord_message;
