+++ src/commands/mod.rs
@@ -5,3 +5,12 @@ pub use ready_announce::Announce;
 mod game_c4;
 mod message_ping;
 mod ready_announce;
+
+impl super::Arbiter {
+	pub fn with_all_commands(mut self) -> Self {
+		self.register_event_handler(Ping::new()).unwrap();
+		self.register_event_handler(Announce).unwrap();
+		self.register_event_handler(ConnectFourDiscord::new()).unwrap();
+		self
+	}
+}
